Create a plan on how to proceed based on Ben's comments
The plan:
I have added the last paragraph in Security page introducing the user to the children pages, which provide details on how the user can further harden security.

So my recommended structure for the pages is as follows:

Security
      Enforcing ASC policy definitions
      Reviewing ASC recommendations

For Enforcing ASC policy definitions:
There is some content added in that page, which needs a lot of polishing. We probably need to refine the exact steps on how the user can:
         1. Configure ASC policy definitions
         2. Remediate non-compliant resources

For Reviewing ASC recommendations:
This page will contain the information under security recommendations listed under “Restrict unauthorised network access“ in Azure Security Center Recommendation.
Ben’s comments to my queries will be a great help.